+++ src/synapsis/synapse/synapse.py
@@ -8,30 +8,52 @@ from ..core.exceptions import LoginError
 
 
 class Synapse(synapseclient.Synapse):
-    __SYNAPSE_INIT_ARGS_DEFAULT__: t.ClassVar[t.Final[dict]] = {'skip_checks': True, 'silent': True}
-    __SYNAPSE_LOGIN_ARGS_DEFAULT__: t.ClassVar[t.Final[dict]] = {'silent': True, 'rememberMe': False, 'forced': True}
+    __SYNAPSE_INIT_ARGS_DEFAULT__: t.ClassVar[t.Final[dict]] = {
+        'skip_checks': True,
+        'silent': True,
+        "multi_threaded": True
+    }
+    __SYNAPSE_LOGIN_ARGS_DEFAULT__: t.ClassVar[t.Final[dict]] = {
+        'silent': True,
+        'rememberMe': False,
+        'forced': True
+    }
+    __CONFIG_DEFAULT__: t.ClassVar[t.Final[dict]] = {
+        "multi_threaded": __SYNAPSE_INIT_ARGS_DEFAULT__['multi_threaded']
+    }
     __synapse_init_args__: dict = {}
     __synapse_login_args__: dict = {}
+    __config__: dict = {}
 
     def __init__(self, **kwargs):
-        init_args = self.__build_init_args__()
-        init_args.update(kwargs)
-        super().__init__(**init_args)
+        self.__init_self__(init_kwargs=kwargs)
+
+    def __init_self__(self, init_kwargs: dict):
+        init_kwargs = self.__build_init_args__(init_kwargs=init_kwargs)
+        super().__init__(**init_kwargs)
+        self.multi_threaded = self.__config__.get('multi_threaded', self.__CONFIG_DEFAULT__['multi_threaded'])
 
     def __configure__(self, synapse_args: dict = {}, **login_args: dict):
         """Sets configuration options for the synapseclient and logs out.
 
-        :param synapse_args: Args for creating the synapsecleint.Synapse instance.
+        :param synapse_args: Args for creating the synapseclient.Synapse instance.
         :param login_args: Args for logging into Synapse.
         :return: None
         """
         self.__synapse_init_args__ = synapse_args or {}
         self.__synapse_login_args__ = login_args or {}
+        self.__config__ = {}
         self.logout()
 
-    def __build_init_args__(self):
-        init_args = {**Synapse.__SYNAPSE_INIT_ARGS_DEFAULT__, **self.__synapse_init_args__}
+    def __build_init_args__(self, init_kwargs: dict):
+        init_args = {**Synapse.__SYNAPSE_INIT_ARGS_DEFAULT__, **init_kwargs}
         self.__from_arg_or_env__(init_args, 'configPath', 'SYNAPSE_CONFIG_FILE')
+
+        # Pull out custom Synapse args.
+        for attr, value in Synapse.__CONFIG_DEFAULT__.items():
+            if attr in init_args:
+                value = init_args.pop(attr)
+            self.__config__[attr] = value
 
         if 'cache_root_dir' not in init_args:
             cache_root_dir = os.path.expandvars(os.path.expanduser((synapseclient.core.cache.CACHE_ROOT_DIR)))
@@ -66,9 +88,8 @@ class Synapse(synapseclient.Synapse):
         :return: None
         :raises LoginError: If logging into Synapse fails.
         """
-        init_args = self.__build_init_args__()
+        self.__init_self__(init_kwargs=self.__synapse_init_args__)
         login_args = self.__build_login_args__()
-        super().__init__(**init_args)
         try:
             self.login(**login_args)
             if hooks: