+++ tests/synapsis/core/test_cli.py
@@ -17,9 +17,10 @@ def test_it_injects_args():
         assert action is not None
 
 
-def test_it_configures_from_args(syn_test_credentials, syn_test_auth_token, synapse_test_config,
+def test_it_configures_from_args(test_credentials, other_test_credentials, synapse_test_config,
                                  clear_env_vars, set_env_vars):
-    test_username, test_password = syn_test_credentials
+    test_username, test_password, test_auth_token = test_credentials
+    other_test_username, other_test_password, other_test_auth_token = other_test_credentials
 
     def reset():
         clear_env_vars()
@@ -38,7 +39,7 @@ def test_it_configures_from_args(syn_test_credentials, syn_test_auth_token, syna
     ###########################################################################
     # User/Pass
     ###########################################################################
-    args = parser.parse_args(['--username', test_username, '--password', test_password])
+    args = parser.parse_args(['--username', other_test_username, '--password', other_test_password])
     cli.configure(args)
     assert Synapsis.logged_in() is False
     Synapsis.login()
@@ -50,7 +51,7 @@ def test_it_configures_from_args(syn_test_credentials, syn_test_auth_token, syna
 
     # ENV Vars
     reset()
-    set_env_vars(username=test_username, password=test_password)
+    set_env_vars(username=other_test_username, password=other_test_password)
     args = parser.parse_args([])
     cli.configure(args)
     assert Synapsis.logged_in() is False
@@ -58,7 +59,7 @@ def test_it_configures_from_args(syn_test_credentials, syn_test_auth_token, syna
     assert Synapsis.logged_in() is True
 
     reset()
-    set_env_vars(username=test_username, password=test_password)
+    set_env_vars(username=other_test_username, password=other_test_password)
     cli.configure(args, login=True)
     assert Synapsis.logged_in() is True
 
@@ -66,7 +67,7 @@ def test_it_configures_from_args(syn_test_credentials, syn_test_auth_token, syna
     # Auth Token
     ###########################################################################
     reset()
-    args = parser.parse_args(['--auth-token', syn_test_auth_token])
+    args = parser.parse_args(['--auth-token', test_auth_token])
     cli.configure(args)
     assert Synapsis.logged_in() is False
     Synapsis.login()
@@ -78,7 +79,7 @@ def test_it_configures_from_args(syn_test_credentials, syn_test_auth_token, syna
 
     # ENV Vars
     reset()
-    set_env_vars(auth_token=syn_test_auth_token)
+    set_env_vars(auth_token=test_auth_token)
     args = parser.parse_args([])
     cli.configure(args)
     assert Synapsis.logged_in() is False
@@ -86,7 +87,7 @@ def test_it_configures_from_args(syn_test_credentials, syn_test_auth_token, syna
     assert Synapsis.logged_in() is True
 
     reset()
-    set_env_vars(auth_token=syn_test_auth_token)
+    set_env_vars(auth_token=test_auth_token)
     cli.configure(args, login=True)
     assert Synapsis.logged_in() is True
 
@@ -94,7 +95,8 @@ def test_it_configures_from_args(syn_test_credentials, syn_test_auth_token, syna
     # Synapse Config File
     ###########################################################################
     reset()
-    config_path = synapse_test_config(username=test_username, password=test_password, auth_token=syn_test_auth_token)
+    config_path = synapse_test_config(username=other_test_username, password=other_test_password,
+                                      auth_token=other_test_auth_token)
     args = parser.parse_args(['--synapse-config', config_path])
     cli.configure(args)
     assert Synapsis.logged_in() is False